moved getLeaves from react to slate core package and added some basic tests
