+++ packages/slate/src/interfaces/text.ts
@@ -1,5 +1,5 @@
 import isPlainObject from 'is-plain-object'
-import { Path } from '..'
+import { Range } from '..'
 
 /**
  * `Text` objects represent the nodes that contain the actual text content of a
@@ -82,5 +82,78 @@ export const Text = {
     }
 
     return true
+  },
+
+  /**
+   * Get the leaves for a text node given decorations.
+   */
+
+  decorations(node: Text, decorations: Range[]): Text[] {
+    let leaves: Text[] = [{ ...node }]
+
+    for (const dec of decorations) {
+      const { anchor, focus, ...rest } = dec
+      const [start, end] = Range.edges(dec)
+      const next = []
+      let o = 0
+
+      for (const leaf of leaves) {
+        const { length } = leaf.text
+        const offset = o
+        o += length
+
+        // If the range encompases the entire leaf, add the range.
+        if (start.offset <= offset && end.offset >= offset + length) {
+          Object.assign(leaf, rest)
+          next.push(leaf)
+          continue
+        }
+
+        // If the range starts after the leaf, or ends before it, continue.
+        if (
+          start.offset > offset + length ||
+          end.offset < offset ||
+          (end.offset === offset && offset !== 0)
+        ) {
+          next.push(leaf)
+          continue
+        }
+
+        // Otherwise we need to split the leaf, at the start, end, or both,
+        // and add the range to the middle intersecting section. Do the end
+        // split first since we don't need to update the offset that way.
+        let middle = leaf
+        let before
+        let after
+
+        if (end.offset < offset + length) {
+          const off = end.offset - offset
+          after = { ...middle, text: middle.text.slice(off) }
+          middle = { ...middle, text: middle.text.slice(0, off) }
+        }
+
+        if (start.offset > offset) {
+          const off = start.offset - offset
+          before = { ...middle, text: middle.text.slice(0, off) }
+          middle = { ...middle, text: middle.text.slice(off) }
+        }
+
+        Object.assign(middle, rest)
+
+        if (before) {
+          next.push(before)
+        }
+
+        next.push(middle)
+
+        if (after) {
+          next.push(after)
+        }
+      }
+
+      leaves = next
+    }
+
+    return leaves
   },
 }
